Converted project to target netstandard2.0 and net46 frameworks. Made changes to support latest EventStore.ClientAPI, EventStore.ClientAPI.NetCore and MassTransit packages.

+++ src/MassTransit.EventStoreIntegration/Saga/EventSourcedSaga.cs
@@ -7,7 +7,7 @@ namespace MassTransit.EventStoreIntegration.Saga
     public abstract class EventSourcedSagaInstance : IEventSourcedSaga
     {
         public Guid CorrelationId { get; set; }
-        public int ExpectedVersion { get; set; }
+        public long ExpectedVersion { get; set; }
 
         public string StreamName =>
             TypeMapping.GetTypeName(GetType()) + "-" + CorrelationId.ToString("N");
@@ -61,12 +61,12 @@ namespace MassTransit.EventStoreIntegration.Saga
             Record(@event);
         }
 
-        /// <summary>
-        /// Initializes this instance using the specified events.
-        /// </summary>
-        /// <param name="events">The events to initialize with.</param>
-        /// <exception cref="System.ArgumentNullException">Thrown when the <paramref name="events"/> are null.</exception>
-        public void Initialize(IEnumerable<object> events)
+		/// <summary>
+		/// Initializes this instance using the specified events.
+		/// </summary>
+		/// <param name="events">The events to initialize with.</param>
+		/// <exception cref="System.ArgumentNullException">Thrown when the <paramref name="events"/> are null.</exception>
+		public void Initialize(IEnumerable<object> events)
         {
             if (events == null) throw new ArgumentNullException(nameof(events));
             if (HasChanges())
@@ -99,7 +99,7 @@ namespace MassTransit.EventStoreIntegration.Saga
 
         void Play(object @event) => _router.Route(@event);
 
-        void Record(object @event) => _recorder.Record(@event);
+		void Record(object @event) => _recorder.Record(@event);
 
         public class SagaInstanceTransitioned
         {
